P1.3a: Split registry handler into thin HTTP files + admin service (#35)
internal/handlers/registry.go was 934 LOC mixing request decode/encode
with YAML/ConfigMap mutation, file storage, and map->model conversion.
That mix blocks Phase 2 (HTTP API -> controller-runtime shim) and made
the handler untestable without infra.

Restructuring (no behavior change, no API change):

- New pkg/services/registry/admin sub-package owns ReloadFromConfig,
  UpdateConfigMap, and UploadLocalMCP. Lives in a sub-package to keep
  the import cycle with internal/handlers broken (same workaround used
  for loader/ in P1.2).
- New Store interface in pkg/services/registry/loader exposes the
  minimal ListMCPServers/CreateMCPServer/GetMCPServer surface the
  admin service needs.
- internal/handlers/registry.go shrinks from 934 to 287 LOC: only the
  struct, constructor, DetectServerType, CRUD, ListMCPServersFiltered,
  and BrowseRegistry remain.
- New internal/handlers/registry_upload.go (167 LOC) holds the three
  upload handlers + isValidMCPFile + generateID.
- New internal/handlers/registry_reload.go (31 LOC) is now a thin
  delegate to AdminService.ReloadFromConfig.
- bootstrap.go wires the admin service before constructing the
  registry handler.

Dead-code removed: convertToMCPServer (152 LOC, no callers anywhere
in the repo), validateMCPServer and enumerateTools (TODO stubs with
no callers).

Out of scope (flagged for follow-up):
- pkg/services/registry/service.go (455 LOC standalone HTTP server
  used by cmd/registry/main.go) — source of the import-cycle
  constraint; likely Phase-2 cleanup.
- Unifying admin's URL-path parser with loader.ParseAndUploadRegistryYAML
  (different input schemas).

Verification: go build ./..., go vet on changed packages, loader tests
pass, no new import cycles, LOC target met.

Closes #4
